+++ app/src/main/java/io/github/sds100/keymapper/UseCases.kt
@@ -195,6 +195,7 @@ object UseCases {
         ServiceLocator.inputMethodAdapter(ctx),
         ServiceLocator.systemFeatureAdapter(ctx),
         ServiceLocator.cameraAdapter(ctx),
+        ServiceLocator.permissionAdapter(ctx),
     )
 
     private fun keyMapperImeMessenger(