+++ app/src/main/java/io/github/sds100/keymapper/actions/ChooseActionViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import io.github.sds100.keymapper.R
 import io.github.sds100.keymapper.system.permissions.Permission
+import io.github.sds100.keymapper.util.Error
 import io.github.sds100.keymapper.util.State
 import io.github.sds100.keymapper.util.containsQuery
 import io.github.sds100.keymapper.util.getFullMessage
@@ -99,14 +100,26 @@ class ChooseActionViewModel(
     }
 
     private fun buildListGroups(): List<SimpleListItemGroup> = buildList {
-        CATEGORY_ORDER.forEach { category ->
+        val listItems = buildListItems(ActionId.entries)
 
+        for (category in CATEGORY_ORDER) {
             val header = getString(ActionUtils.getCategoryLabel(category))
 
-            val actionIds = ActionId.entries.filter { ActionUtils.getCategory(it) == category }
-
-            add(SimpleListItemGroup(header, buildListItems(actionIds)))
+            val group = SimpleListItemGroup(
+                header,
+                items = listItems.filter {
+                    it.isEnabled &&
+                        ActionUtils.getCategory(ActionId.valueOf(it.id)) == category
+                },
+            )
+            add(group)
         }
+
+        val unsupportedGroup = SimpleListItemGroup(
+            header = getString(R.string.choose_action_group_unsupported),
+            items = listItems.filter { !it.isEnabled },
+        )
+        add(unsupportedGroup)
     }
 
     private fun buildListItems(
@@ -119,12 +132,9 @@ class ChooseActionViewModel(
             val title = getString(ActionUtils.getTitle(actionId))
             val icon = ActionUtils.getComposeIcon(actionId)
 
-            val requiresRoot = ActionUtils.getRequiredPermissions(actionId)
-                .contains(Permission.ROOT)
-
             val subtitle = when {
+                error == Error.PermissionDenied(Permission.ROOT) -> getString(R.string.choose_action_warning_requires_root)
                 error != null -> error.getFullMessage(this@ChooseActionViewModel)
-                requiresRoot -> getString(R.string.choose_action_warning_requires_root)
                 else -> null
             }